requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.34,<1.0
scikit-learn>=1.8,<2.0
numpy>=2.3,<3.0
pydantic-settings>=2.8,<3.0
python-dotenv>=1.0,<2.0
httpx>=0.28,<1.0
